level_editor: Support dragging players around
Only render jump paths when cursor is near a player.
Jump paths are rather expensive to render, as they run the equivalent
of full-level collision check and physics tick as many times as
necessary to get the player to reach ground or fall off the screen.

Also, the paths aren't really useful when you're not focusing on the
players, so might as well not waste the CPU cycles when not prompted to.