collisions: Fix tiles colliding when put back between others
As tiles were claiming their edges too,
modifying the exclusion criteria for possible collidable tiles to
discard tiles that may touch this one on an axis solved that issue.

There was another issue with the level editor:
Some tiles would get an extra pixel size.
Turns out that was a forgotten grid-size multiplier.